# specs/001-rag-chatbot-auth/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Physical AI & Humanoid Robotics RAG Chatbot API
  description: Backend API for authenticated RAG chatbot with conversation history, recommendations, and reading progress tracking
  version: 1.0.0
  contact:
    name: API Support
    email: support@physical-ai-book.com
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.physical-ai-book.com
    description: Production server (Hugging Face Spaces)
...
security:
  - bearerAuth: []
paths:
...
  /chat/global:
    post:
      tags:
        - Chat
      summary: Ask question about book (global search)
      description: Query across all book chapters using RAG retrieval
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                  minLength: 3
                  example: What is ROS 2 and how does it differ from ROS 1?
                conversation_id:
                  type: string
                  format: uuid
                  description: Optional conversation ID for multi-turn context
      responses:
        '200':
          description: Answer generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          description: Invalid query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          description: Daily quota exceeded (1200 requests/day)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /chat/local:
    post:
      tags:
        - Chat
      summary: Ask question about selected text
      description: Query using only user-selected text as context (strict isolation)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
                - file_path
                - chunk_indices
              properties:
                query:
                  type: string
                  minLength: 3
                  example: What simulation capabilities are mentioned here?
                file_path:
                  type: string
                  example: docs/module-2-simulation/isaac-sim.md
                chunk_indices:
                  type: array
                  items:
                    type: integer
                  minItems: 1
                  example: [0, 1, 2]
                conversation_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Answer generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          description: Invalid query or selection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # Conversation History Endpoints
...
  /progress:
    get:
      tags:
        - Progress
      summary: Get reading progress
      description: Retrieve user's progress across all chapters
      responses:
        '200':
          description: Progress retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  progress:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReadingProgress'
                  total_chapters:
                    type: integer
                  completed_chapters:
                    type: integer
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Progress
      summary: Update reading progress
      description: Update progress for specific chapter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - chapter_id
                - completion_percentage
              properties:
                chapter_id:
                  type: string
                  example: module-1-ros2/introduction
                completion_percentage:
                  type: integer
                  minimum: 0
                  maximum: 100
                  example: 75
                time_spent_seconds:
                  type: integer
                  minimum: 0
                  example: 120
      responses:
        '200':
          description: Progress updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadingProgress'
        '400':
          description: Invalid progress data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # Internal Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminAuth:
      type: apiKey
      in: header
      name: X-Admin-Key
  responses:
    UnauthorizedError:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    ChatResponse:
      type: object
      properties:
        answer:
          type: string
          description: Generated answer from book content
        sources:
          type: array
          items:
            $ref: '#/components/schemas/Source'
        conversation_id:
          type: string
          format: uuid
          description: Conversation ID for multi-turn context
        query_mode:
          type: string
          enum: [global, local]
    Source:
      type: object
      properties:
        file_path:
          type: string
        section_heading:
          type: string
        chunk_index:
          type: integer
        similarity_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
...
    ReadingProgress:
      type: object
      properties:
        chapter_id:
          type: string
        chapter_title:
          type: string
        completion_percentage:
          type: integer
          minimum: 0
          maximum: 100
        time_spent_seconds:
          type: integer
        last_accessed:
          type: string
          format: date-time
        completed:
          type: boolean
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties: true